+++ uninstall.sh
@@ -1,0 +1,2 @@
+#!/system/bin/sh
+rm -Rf /data/adb/gki_patcher_curl_ui